requirements.txt
python-dotenv>=1.0.0     # .env 配置
requests>=2.31.0         # HTTP 请求
fake-useragent>=1.4.0    # User-Agent 轮换
tenacity>=8.2.0          # 重试机制
sqlalchemy>=2.0.0        # ORM (SQLite)
pandas>=2.0.0            # 数据处理
numpy>=1.24.0            # 数值计算
akshare>=1.12.0          # 备用数据源
flask>=3.0.0             # Web 页面
markdown>=3.4.0          # 文档 Markdown 渲染
flask-limiter>=3.5.0     # API 速率限制
